requirements.txt
Flask==3.0.0
flask-cors==4.0.0
openai==1.52.0
google-cloud-texttospeech==2.16.3
python-dotenv==1.0.0
gunicorn==21.2.0
httpx==0.27.2
